+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file.
 
+## WIP
+
+- Enabling/disabling the module will now properly update the profiler config in app/etc/env.php
+
 ## [5.1.1] - 2021-08-05
 [5.1.1]: https://github.com/Smile-SA/magento2-module-debug-toolbar/compare/5.1.0...5.1.1